+++ resources/lib/webservice.py
@@ -27,21 +27,26 @@ from http.server import HTTPServer, BaseHTTPRequestHandler
 from http.client import HTTPConnection
 
 # AKL modules
+from akl import settings
 from resources.lib import globals, apiqueries
 from resources.lib.commands import api_commands
 
 logger = logging.getLogger(__name__)
 
+
 #################################################################################################
 class WebService(threading.Thread):
-    
-    HOST = globals.WEBSERVER_HOST
-    PORT = globals.WEBSERVER_PORT
     
     ''' Run a webservice for api communication.
     '''
     def __init__(self):
         self.server = None
+        self.port = settings.getSettingAsInt('webserver_port')
+        self.host = globals.WEBSERVER_HOST
+        
+        if self.port is None or self.port == 0:
+            self.port = globals.WEBSERVER_PORT
+        
         threading.Thread.__init__(self)
 
     def is_alive(self):
@@ -52,7 +57,7 @@ class WebService(threading.Thread):
         s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
         try:
-            s.connect((WebService.HOST, WebService.PORT))
+            s.connect((self.host, self.port))
             s.sendall("")
         except Exception as error:
             logger.fatal(f'Exception in webservice.is_alive {str(error)}')
@@ -72,8 +77,8 @@ class WebService(threading.Thread):
         ''' Called when the thread needs to stop
         '''
         try:
-            logger.info(f"Stopping AKL webservice({WebService.HOST}:{WebService.PORT})")
-            conn = HTTPConnection(f"{WebService.HOST}:{WebService.PORT}")
+            logger.info(f"Stopping AKL webservice({self.host}:{self.port})")
+            conn = HTTPConnection(f"{self.host}:{self.port}")
             conn.request("QUIT", "/")
             conn.getresponse()
         except Exception as error:
@@ -87,8 +92,8 @@ class WebService(threading.Thread):
         '''
         self.stop(check_alive=True)
 
-        logger.info("Startup AKL webservice({}:{})".format(WebService.HOST, WebService.PORT))
-        server = AelHttpServer((WebService.HOST, WebService.PORT), RequestHandler)
+        logger.info("Startup AKL webservice({}:{})".format(self.host, self.port))
+        server = AelHttpServer((self.host, self.port), RequestHandler)
         
         try:
             server.serve_forever()